fix(logs): restore canonical log URLs
