+++ src/main.ts
@@ -160,6 +160,17 @@ const router = createRouter({
     { path: '/dashboard/settings/organization/plans', redirect: '/settings/organization/plans' },
     { path: '/dashboard/settings/organization/usage', redirect: '/settings/organization/usage' },
     { path: '/p/:package', redirect: to => `/app/${(to.params as { package: string }).package}` },
+    {
+      path: '/p/:package/logs',
+      redirect: (to) => {
+        const { tab: _, ...query } = to.query
+        return {
+          path: `/app/${encodeURIComponent(String((to.params as { package: string }).package))}/logs`,
+          query,
+          hash: to.hash,
+        }
+      },
+    },
     { path: '/dashboard/apikeys', redirect: '/apikeys' },
     { path: '/dashboard/settings/account', redirect: '/settings/account' },
     { path: '/dashboard/settings/change-password', redirect: '/settings/account/change-password' },
@@ -168,11 +179,32 @@ const router = createRouter({
     { path: '/dashboard/settings/organization/members', redirect: '/settings/organization/members' },
     { path: '/dashboard/settings/organization/plans', redirect: '/settings/organization/plans' },
     { path: '/dashboard/settings/organization/usage', redirect: '/settings/organization/usage' },
+    {
+      path: '/app/:app/logs/raw',
+      redirect: (to) => {
+        const { tab: _, ...query } = to.query
+        return {
+          path: `/app/${encodeURIComponent(String((to.params as { app: string }).app))}/logs`,
+          query,
+          hash: to.hash,
+        }
+      },
+    },
     { path: '/app/p/:package', redirect: to => `/app/${(to.params as { package: string }).package}` },
     { path: '/app/p/:package/bundles', redirect: to => `/app/${(to.params as { package: string }).package}` },
     { path: '/app/p/:package/channels', redirect: to => `/app/${(to.params as { package: string }).package}` },
     { path: '/app/p/:package/devices', redirect: to => `/app/${(to.params as { package: string }).package}` },
-    { path: '/app/p/:package/logs', redirect: to => `/app/${(to.params as { package: string }).package}` },
+    {
+      path: '/app/p/:package/logs',
+      redirect: (to) => {
+        const { tab: _, ...query } = to.query
+        return {
+          path: `/app/${encodeURIComponent(String((to.params as { package: string }).package))}/logs`,
+          query,
+          hash: to.hash,
+        }
+      },
+    },
     { path: '/app/package/:package', redirect: to => `/app/${(to.params as { package: string }).package}` },
     { path: '/app/package/:package/settings', redirect: to => `/app/${(to.params as { package: string }).package}` },
     ...setupLayouts(newRoutes),
